requirements.txt
asgiref==3.3.1
Django==3.1.3
djangorestframework==3.12.4
numpy==1.19.4
pandas==1.1.4
python-dateutil==2.8.1
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
